+++ packages/sycamore-macro/src/lib.rs
@@ -10,7 +10,6 @@ use syn::{parse_macro_input, DeriveInput};
 mod component;
 mod props;
 mod view;
-mod view_old;
 
 /// A macro for ergonomically creating complex UI complex layouts.
 ///
