+++ packages/sycamore-web/Cargo.toml
@@ -1,40 +1,31 @@
 [package]
 name = "sycamore-web"
-categories = ["gui", "wasm", "web-programming"]
-description = "Web renderer for the Sycamore"
+version = "0.9.0-beta.2"
 edition = "2021"
-homepage = "https://github.com/sycamore-rs/sycamore"
-keywords = ["wasm", "gui", "reactive", "web"]
-license = "MIT"
-readme = "../../README.md"
-repository = "https://github.com/sycamore-rs/sycamore"
-version.workspace = true
 
 [dependencies]
-hashbrown = "0.14.1"
-html-escape = "0.2.12"
-indexmap = "2.0.2"
-js-sys = "0.3.60"
-once_cell = "1.16.0"
+html-escape = { version = "0.2.13", optional = true }
+js-sys = "0.3.67"
+paste = "1.0.14"
+once_cell = "1.19.0"
 sycamore-core = { path = "../sycamore-core", version = "0.9.0-beta.2" }
+sycamore-macro = { path = "../sycamore-macro", version = "0.9.0-beta.2" }
 sycamore-reactive = { path = "../sycamore-reactive", version = "0.9.0-beta.2" }
-wasm-bindgen = "0.2.83"
-web-sys = { version = "0.3.60", features = [
-	"console",
+wasm-bindgen = "0.2.92"
+web-sys = { version = "0.3.69", features = [
 	"Comment",
-	"Document",
-	"DocumentFragment",
-	"DomTokenList",
-	"Element",
-	"HtmlCollection",
-	"HtmlElement",
-	"HtmlInputElement",
-	"HtmlTemplateElement",
+	"console",
 	"Node",
 	"NodeList",
-	"Text",
 	"Window",
+	"Document",
+	"Element",
+	"EventListener",
+	"HtmlElement",
+	"Text",
+
 	# Event types
+	"Event",
 	"AnimationEvent",
 	"BeforeUnloadEvent",
 	"CompositionEvent",
@@ -64,8 +55,10 @@ web-sys = { version = "0.3.60", features = [
 ] }
 
 [dev-dependencies]
-sycamore = { path = "../sycamore", features = ["ssr"] }
+expect-test = "1.4.1"
 
 [features]
-hydrate = []
-ssr = ["hydrate"]
+default = ["dom", "wasm-bindgen-interning"]
+dom = []
+ssr = ["dep:html-escape"]
+wasm-bindgen-interning = ["wasm-bindgen/enable-interning"]
