--- packages/sycamore-web2/Cargo.toml
@@ -1,64 +1,0 @@
-[package]
-name = "sycamore-web2"
-version = "0.9.0-beta.2"
-edition = "2021"
-
-[dependencies]
-html-escape = { version = "0.2.13", optional = true }
-js-sys = "0.3.67"
-paste = "1.0.14"
-once_cell = "1.19.0"
-sycamore-core = { path = "../sycamore-core", version = "0.9.0-beta.2" }
-sycamore-macro = { path = "../sycamore-macro", version = "0.9.0-beta.2" }
-sycamore-reactive = { path = "../sycamore-reactive", version = "0.9.0-beta.2" }
-wasm-bindgen = "0.2.92"
-web-sys = { version = "0.3.69", features = [
-	"Comment",
-	"console",
-	"Node",
-	"NodeList",
-	"Window",
-	"Document",
-	"Element",
-	"EventListener",
-	"HtmlElement",
-	"Text",
-
-	# Event types
-	"Event",
-	"AnimationEvent",
-	"BeforeUnloadEvent",
-	"CompositionEvent",
-	"DeviceMotionEvent",
-	"DeviceOrientationEvent",
-	"DragEvent",
-	"ErrorEvent",
-	"Event",
-	"FocusEvent",
-	"GamepadEvent",
-	"HashChangeEvent",
-	"InputEvent",
-	"KeyboardEvent",
-	"MessageEvent",
-	"MouseEvent",
-	"PageTransitionEvent",
-	"PointerEvent",
-	"PopStateEvent",
-	"ProgressEvent",
-	"PromiseRejectionEvent",
-	"SecurityPolicyViolationEvent",
-	"StorageEvent",
-	"SubmitEvent",
-	"TouchEvent",
-	"TransitionEvent",
-	"WheelEvent",
-] }
-
-[dev-dependencies]
-expect-test = "1.4.1"
-
-[features]
-default = ["dom", "wasm-bindgen-interning"]
-dom = []
-ssr = ["dep:html-escape"]
-wasm-bindgen-interning = ["wasm-bindgen/enable-interning"]
