--- packages/sycamore-web2/examples/js-framework-benchmark/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "js-framework-benchmark"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-console_error_panic_hook = "0.1.7"
-getrandom = { version = "0.2.8", features = ["js"] }
-rand = "0.8.5"
-sycamore-web2 = { path = "../../" }
-web-sys = "0.3.60"
