--- packages/sycamore-web2/examples/js-framework-benchmark/index.css
@@ -1,27 +1,0 @@
-body {
-  padding: 10px 0 0 0;
-  margin: 0;
-  overflow-y: scroll;
-}
-#duration {
-  padding-top: 0px;
-}
-.jumbotron {
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-.test-data a {
-  display: block;
-}
-.preloadicon {
-  position: absolute;
-  top: -20px;
-  left: -20px;
-}
-.col-sm-6.smallpad {
-  padding: 5px;
-}
-.jumbotron .row h1 {
-  font-size: 40px;
-}
-
