--- packages/sycamore-web2/examples/js-framework-benchmark/index.html
@@ -1,18 +1,0 @@
-<!doctype html>
-<html>
-
-<head>
-	<meta charset="utf-8" />
-	<title>Sycamore Beta</title>
-	<meta name="viewport" content="width=device-width, initial-scale=1" />
-	<link data-trunk rel="css" href="index.css" />
-	<link data-trunk rel="css" href="bootstrap.min.css" />
-	<link data-trunk rel="rust" data-wasm-opt="4" data-reference-types data-weak-refs />
-</head>
-
-<body>
-	<span class="preloadicon glyphicon glyphicon-remove" aria-hidden="true"></span>
-	<div id="main"></div>
-</body>
-
-</html>
