--- packages/sycamore-web2/examples/simple/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "simple"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-console_error_panic_hook = "0.1.7"
-sycamore-web2 = { path = "../../", features = ["dom", "ssr"] }
-web-sys = "0.3.67"
