--- packages/sycamore-web2/examples/simple/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <link
-      data-trunk
-      rel="rust"
-      data-wasm-opt="4"
-      data-reference-types
-      data-weak-refs
-    />
-    <link data-trunk rel="tailwind-css" href="index.scss" />
-  </head>
-
-  <body>
-    DOM render test:
-    <div id="container"></div>
-    Hydrate render test:
-    <div id="hydrate-test"></div>
-  </body>
-</html>
