--- packages/sycamore-web2/examples/simple/index.scss
@@ -1,3 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
