--- packages/sycamore-web2/examples/simple/tailwind.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-	mode: "jit",
-	content: {
-		files: ["./src/**/*.rs", "index.html"],
-	},
-	darkMode: "media",
-	plugins: [],
-};
